requirements.txt
flytekit
flytekitplugins-domino


